+++ src/components/searcher/contents.rs
@@ -49,13 +49,31 @@ mod contents {
                     CrosstermEvent::Key(KeyEvent {
                         code: KeyCode::Char('j'),
                         modifiers: KeyModifiers::NONE,
-                        ..
                     }) => Some(Action::Down),
+                    CrosstermEvent::Key(KeyEvent {
+                        code: KeyCode::Char('J'),
+                        modifiers: KeyModifiers::SHIFT,
+                    }) => Some(Action::ReallyDown),
+                    CrosstermEvent::Key(KeyEvent {
+                        code: KeyCode::Char('j'),
+                        modifiers: KeyModifiers::CONTROL,
+                    }) => Some(Action::ScrollDown),
                     CrosstermEvent::Key(KeyEvent {
                         code: KeyCode::Char('k'),
                         modifiers: KeyModifiers::NONE,
-                        ..
                     }) => Some(Action::Up),
+                    CrosstermEvent::Key(KeyEvent {
+                        code: KeyCode::Char('K'),
+                        modifiers: KeyModifiers::SHIFT,
+                    }) => Some(Action::ReallyUp),
+                    CrosstermEvent::Key(KeyEvent {
+                        code: KeyCode::Char('k'),
+                        modifiers: KeyModifiers::CONTROL,
+                    }) => Some(Action::ScrollUp),
+                    CrosstermEvent::Key(KeyEvent {
+                        code: KeyCode::Char('r'),
+                        modifiers: KeyModifiers::NONE,
+                    }) => Some(Action::Refresh),
                     CrosstermEvent::Key(KeyEvent {
                         code: KeyCode::Char('l'),
                         ..
@@ -64,6 +82,14 @@ mod contents {
                         code: KeyCode::Enter,
                         ..
                     }) => Some(Action::Edit),
+                    CrosstermEvent::Key(KeyEvent {
+                        code: KeyCode::Char('g'),
+                        modifiers: KeyModifiers::NONE,
+                    }) => Some(Action::Goto),
+                    CrosstermEvent::Key(KeyEvent {
+                        code: KeyCode::Char('G'),
+                        modifiers: KeyModifiers::SHIFT,
+                    }) => Some(Action::ReallyGoto),
                     CrosstermEvent::Key(KeyEvent {
                         code: KeyCode::Char('y'),
                         modifiers: KeyModifiers::NONE,
@@ -199,20 +225,20 @@ pub use event::Event;
 
 mod state {
     use super::{Action, Effect, Props};
+    use crate::clipboard::Clipboard;
     use crate::phrase_searcher::{FileHit, LineHit, PhraseSearcher};
     use crate::programs::{VimArgs, VimArgsBuilder};
     use crate::rendering::Size;
     use crate::Stateful;
 
     use std::cmp::Ordering;
-    use std::path::MAIN_SEPARATOR as PATH_SEPARATOR;
-    use std::path::{Path, PathBuf};
+    use std::path::{Path, PathBuf, MAIN_SEPARATOR as PATH_SEPARATOR};
 
-    use copypasta::{ClipboardContext as Clipboard, ClipboardProvider};
-
+    #[derive(Debug, PartialEq, Eq, Default)]
     pub struct State {
         size: Size,
         directory: PathBuf,
+        phrase: Option<String>,
         focussed: bool,
         searched: bool,
         hits: Vec<FileHit>,
@@ -227,6 +253,7 @@ mod state {
             Self {
                 size: props.size,
                 directory: props.directory,
+                phrase: None,
                 focussed: false,
                 searched: false,
                 hits: Vec::new(),
@@ -367,10 +394,11 @@ mod state {
             Some(Effect::Unfocus)
         }
 
-        fn search(&mut self, phrase: String) -> Option<Effect> {
+        fn search(&mut self, phrase: &str) -> Option<Effect> {
             self.focus();
+            self.phrase = Some(phrase.to_string());
 
-            let phrase_searcher = PhraseSearcher::new(&self.directory, &phrase);
+            let phrase_searcher = PhraseSearcher::new(&self.directory, phrase);
             self.hits = phrase_searcher.collect();
             self.searched = true;
 
@@ -408,7 +436,33 @@ mod state {
             None
         }
 
-        fn scroll_down(&mut self, rows: usize) {
+        /// Select the last file hit and adjust the scroll if necessary.
+        fn really_down(&mut self) -> Option<Effect> {
+            if self.hits.is_empty() {
+                return None;
+            }
+
+            self.file_offset = self.hits.len() - 1;
+            self.line_offset = None;
+            self.file_selected = 0;
+            self.line_selected = None;
+
+            let up_adjustment: usize;
+            {
+                let last_file_hit: &FileHit = self.hits.last().unwrap();
+                let number_of_line_hits: usize = last_file_hit.line_hits().len();
+                up_adjustment = self.size.rows - (number_of_line_hits + 1);
+            }
+            // For now, scroll up one line at a time b/c there seems to be a bug w/ scrolling too
+            // many lines at a time
+            for _ in 0..up_adjustment {
+                self.scroll_up(1);
+            }
+
+            None
+        }
+
+        fn scroll_down(&mut self, rows: usize) -> Option<Effect> {
             for _ in 0..rows {
                 match self.line_offset {
                     None => {
@@ -431,6 +485,7 @@ mod state {
                     }
                 }
             }
+            None
         }
 
         fn up(&mut self) -> Option<Effect> {
@@ -496,7 +551,21 @@ mod state {
             None
         }
 
-        fn scroll_up(&mut self, mut rows: usize) {
+        /// Select the first file hit and adjust the scroll position if necessary.
+        fn really_up(&mut self) -> Option<Effect> {
+            if self.hits.is_empty() {
+                return None;
+            }
+
+            self.file_offset = 0;
+            self.line_offset = None;
+            self.file_selected = 0;
+            self.line_selected = None;
+
+            None
+        }
+
+        fn scroll_up(&mut self, mut rows: usize) -> Option<Effect> {
             while rows > 0 {
                 match self.line_offset {
                     Some(line_offset) => {
@@ -512,6 +581,11 @@ mod state {
 
                         if rows == line_offset + 1 {
                             self.line_offset = None;
+                            if self.file_selected == 0 {
+                                if let Some(line_selected) = self.line_selected {
+                                    self.line_selected = Some(line_selected + rows + 1);
+                                }
+                            }
                             break;
                         }
 
@@ -526,7 +600,7 @@ mod state {
                         }
                         rows -= line_offset + 1;
                         self.file_offset -= 1;
-                        self.line_offset = Some(self.hit().unwrap().line_hits().len());
+                        self.line_offset = Some(self.hits[self.file_offset].line_hits().len());
                     }
                     None => {
                         if self.file_offset == 0 {
@@ -535,10 +609,20 @@ mod state {
 
                         rows -= 1;
                         self.file_offset -= 1;
-                        self.line_offset = Some(self.hit().unwrap().line_hits().len());
+                        self.file_selected += 1;
+                        self.line_offset = Some(self.hits[self.file_offset].line_hits().len());
                     }
                 }
             }
+            None
+        }
+
+        /// Refresh the hits by searching for the phrase again.
+        fn refresh(&mut self) -> Option<Effect> {
+            if let Some(phrase) = self.phrase.clone() {
+                return self.search(&phrase);
+            }
+            None
         }
 
         fn edit(&mut self) -> Option<Effect> {
@@ -555,6 +639,29 @@ mod state {
             let vim_args: VimArgs = vim_args_builder.build();
 
             Some(Effect::OpenVim(vim_args))
+        }
+
+        fn goto(&mut self) -> Option<Effect> {
+            self._goto(false)
+        }
+
+        fn really_goto(&mut self) -> Option<Effect> {
+            self._goto(true)
+        }
+
+        fn _goto(&mut self, really: bool) -> Option<Effect> {
+            if let Some(file_hit) = self.hit() {
+                let path: &Path = file_hit.path();
+                let directory = path.parent().unwrap().to_path_buf();
+                let file: Option<PathBuf> = if really {
+                    Some(path.to_path_buf())
+                } else {
+                    None
+                };
+
+                return Some(Effect::Goto { directory, file });
+            }
+            None
         }
 
         /// If a file path is selected, copy it to the system clipboard. Else if the line of a file is selected, then copy it.
@@ -586,8 +693,8 @@ mod state {
                         path
                     }
                 };
-                let mut clipboard = Clipboard::new().unwrap();
-                clipboard.set_contents(contents).unwrap();
+                let mut clipboard = Clipboard::new();
+                clipboard.copy(contents);
             }
             None
         }
@@ -598,13 +705,74 @@ mod state {
             match action {
                 Action::Resize { size } => self.resize(size),
                 Action::Unfocus => self.unfocus(),
-                Action::Search { phrase } => self.search(phrase),
+                Action::Search { phrase } => self.search(&phrase),
                 Action::Down => self.down(),
+                Action::ReallyDown => self.really_down(),
+                Action::ScrollDown => self.scroll_down(1),
                 Action::Up => self.up(),
+                Action::ReallyUp => self.really_up(),
+                Action::ScrollUp => self.scroll_up(1),
+                Action::Refresh => self.refresh(),
                 Action::Edit => self.edit(),
+                Action::Goto => self.goto(),
+                Action::ReallyGoto => self.really_goto(),
                 Action::Yank => self.yank(),
                 Action::ReallyYank => self.really_yank(),
             }
+        }
+    }
+
+    #[cfg(test)]
+    mod tests {
+        use super::*;
+
+        use test_case::test_case;
+
+        #[test_case(&mut State::default(), 0, State::default();)]
+        #[test_case(
+            &mut State{
+                size: Size{rows: 1, columns: 2},
+                hits: vec![FileHit::new(Path::new(""), vec![LineHit::new(0, "")])],
+                ..Default::default()
+            },
+            1,
+            State{
+                size: Size{rows: 1, columns: 2},
+                hits: vec![FileHit::new(Path::new(""), vec![LineHit::new(0, "")])],
+                ..Default::default()
+            };
+        )]
+        #[test_case(
+            &mut State{
+                size: Size{rows: 2, columns: 5},
+                hits: vec![
+                    FileHit::new(Path::new(""), vec![LineHit::new(0, ""), LineHit::new(1, "")]),
+                    FileHit::new(Path::new(""), vec![LineHit::new(0, "")]),
+                ],
+                file_offset: 1,
+                line_offset: None,
+                file_selected: 0,
+                line_selected: None,
+                ..Default::default()
+            },
+            1,
+            State{
+                size: Size{rows: 2, columns: 5},
+                hits: vec![
+                    FileHit::new(Path::new(""), vec![LineHit::new(0, ""), LineHit::new(1, "")]),
+                    FileHit::new(Path::new(""), vec![LineHit::new(0, "")]),
+                ],
+                file_offset: 0,
+                line_offset: Some(2),
+                file_selected: 1,
+                line_selected: None,
+                ..Default::default()
+            };
+        )]
+        fn test_scroll_up(state: &mut State, rows: usize, expected_state: State) {
+            state.scroll_up(rows);
+
+            assert_eq!(*state, expected_state);
         }
     }
 }
@@ -618,8 +786,15 @@ mod action {
         Unfocus,
         Search { phrase: String },
         Down,
+        ReallyDown,
+        ScrollDown,
         Up,
+        ReallyUp,
+        ScrollUp,
+        Refresh,
         Edit,
+        Goto,
+        ReallyGoto,
         Yank,
         ReallyYank,
     }
@@ -629,8 +804,14 @@ use action::Action;
 mod effect {
     use crate::programs::VimArgs;
 
+    use std::path::PathBuf;
+
     pub enum Effect {
         Unfocus,
+        Goto {
+            directory: PathBuf,
+            file: Option<PathBuf>,
+        },
         OpenVim(VimArgs),
     }
 }